Refactor the RecordBatch writing process to eliminate in-memory accumulation. #354
