+++ mosaicod/crates/mosaicod-rw/Cargo.toml
@@ -10,7 +10,6 @@ license.workspace = true
 mosaicod-core = { workspace = true }
 mosaicod-ext = { workspace = true }
 
-log = { workspace = true }
 arrow = { workspace = true }
 parquet = { workspace = true }
 thiserror = { workspace = true }
@@ -19,5 +18,4 @@ tokio = { workspace = true }
 bytes = { workspace = true }
 
 [dev-dependencies]
-rand = { workspace = true }
 arrow-schema = { workspace = true }
