+++ mosaicod/crates/mosaicod-rw/src/lib.rs
@@ -6,13 +6,13 @@ pub use errors::Error;
 pub mod format;
 pub use format::*;
 
-pub mod chunk_writer;
-pub use chunk_writer::{ChunkMetadata, ChunkWriter};
+pub mod chunk_encoder;
+pub use chunk_encoder::{ChunkMetadata, InMemoryChunkEncoder};
 
 mod writer;
 
-pub mod chunked_writer;
-pub use chunked_writer::{ChunkedWriter, ChunkedWriterSummary};
+pub mod chunk_writer;
+pub use chunk_writer::{ChunkWriter, SerializedChunk};
 
 pub mod chunk_reader;
 pub use chunk_reader::ChunkReader;
