+++ mosaicod/crates/mosaicod-rw/src/writer.rs
@@ -8,7 +8,6 @@ use std::sync::Arc;
 
 pub enum Writer {
     /// Parquet file format <https://parquet.apache.org/docs/file-format/>
-    /// (cabba) TODO: evaluate `AsyncArrowWriter`
     Parquet(ArrowWriter<Vec<u8>>),
 }
